four loops prioritizing algorithm
Description of files:
color_obj.py contains the class defining each block or ball
robot_2019.py contains the class describing the state of the robot
robot_plot.py is a simple GUI used for testing

How to use:
The easiest way to test is to use robot_plot.py, which is designed to run in IDLE. After running the file, the robot (called JJ) and objects will be initialized. You have the option to either create random objects or choose specific locations depending on what you comment out in robot_plot.py. To see the field, use show_field(). To see the ordering of goals, use loops_gui(), and then the show_field() command again. To only return a list of goal locations, use the function JJ.four_loops(). Loops_gui() and JJ.four_loops() both execute the same commands on the field, but JJ.four_loops() does not plot the results. Note that every time you run either loops_gui() or JJ.four_loops(), you will need to reset the field using the reset_field() command. 

Important details:
The robot_plot file is not intended to be used in the final program, but is mainly for testing. The four_loops() function in the file robot_2019 is also mostly for testing, since it assumes that the location of all objects is known, and does not allow for interrupts from sensor data. It might be useful as an outline, though.
The most important part of these files is the next_obj() command in robot_2019.py. It will need to be modified to allow for live inputs, but it does controls the order that objects are retrieved in.